chore: move error classes to common
