+++ packages/backend/src/dbt/dbtCloudV2RpcClient.ts
@@ -1,4 +1,4 @@
-import { DbtError, LightdashError, NetworkError } from '../errors';
+import { DbtError, LightdashError, NetworkError } from 'common';
 import { DbtRpcClientBase, DEFAULT_HEADERS } from './dbtRpcClientBase';
 
 export class DbtCloudV2RpcClient extends DbtRpcClientBase {
