+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -1,6 +1,8 @@
 import {
+    AlreadyProcessingError,
     ApiQueryResults,
     ApiSqlQueryResults,
+    AuthorizationError,
     countTotalFilterRules,
     CreateJob,
     CreateProject,
@@ -9,6 +11,7 @@ import {
     ExploreError,
     fieldId,
     FilterableField,
+    ForbiddenError,
     formatRows,
     getDimensions,
     getFields,
@@ -21,6 +24,8 @@ import {
     JobStepType,
     JobType,
     MetricQuery,
+    MissingWarehouseCredentialsError,
+    NotExistsError,
     Project,
     ProjectCatalog,
     SessionUser,
@@ -31,14 +36,7 @@ import {
 } from 'common';
 import { v4 as uuidv4 } from 'uuid';
 import { analytics } from '../../analytics/client';
-import {
-    AlreadyProcessingError,
-    AuthorizationError,
-    errorHandler,
-    ForbiddenError,
-    MissingWarehouseCredentialsError,
-    NotExistsError,
-} from '../../errors';
+import { errorHandler } from '../../errors';
 import Logger from '../../logger';
 import { JobModel } from '../../models/JobModel/JobModel';
 import { OnboardingModel } from '../../models/OnboardingModel/OnboardingModel';
